requirements.txt
fastapi
graphene
graphene-sqlalchemy
uvicorn
psycopg2
python-jose
python-multipart
python-dotenv